requirements.txt
twitchio>=2.6.0
aiosqlite>=0.19.0
aiohttp>=3.9.5
python-dotenv>=1.0.1
qrcode[pil]>=7.4.2
uvloop>=0.19.0; sys_platform == 'linux' or sys_platform == 'darwin'
humanize>=4.9.0
rapidfuzz>=3.9.7
PySide6>=6.7.2
qasync>=0.27.1